pyproject.toml
[project]
name = "brave-orders"
version = "0.0.1"
description = "Order notes"
authors = [
    {name = "gcca", email = "git@gcca.dev"},
]
dependencies = ["django>=6.0", "djangorestframework>=3.16.1"]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "AGPL-3.0-or-later"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[dependency-groups]
dev = [
    "django-stubs>=5.2.8",
]
